Fixed defect 44
This page in particular was loading a lot of information on load. This
seems to have been causeing the issue because promises weren't resolving
apporpriatly.

I moved the promises into the router to resolve before the route even
loads. This has helped the template to bind the values correctly.

See
(https://medium.com/opinionated-angularjs/advanced-routing-and-resolves-a2fcbf874a1c)[https://medium.com/opinionated-angularjs/advanced-routing-and-resolves-a2fcbf874a1c]
for motivation behind this technique.