DE87 LoginForm Menu IDs
The addition of the menu and mobile menu included two new login forms
that did not have their own unique id's.

To solve this I modified the login-form directive to check for an
attribute called prefix that when set will override the passwordPrefix
that is used in the controllers of the other login forms. It can easily
be used like :

```
<login-form prefix="whatever-you-want"> </login-form>
```