DE180 submit filter dates on enter key
Since the form does not have a submit button, I had to add a hidden
input type. The problem is that if you add an input of type 'submit',
you can't also create it hidden. To remedy this, I added the
`display:none` style to the input tag. I think the best solution in the
long run is to actually add a submit button for the user to press.