Fix browser-sync-dev
With the performance changes, browser-sync-dev and webpack-dev-server were not watching for changes in core anymore.  This was due to some consolidation of htmlreplace code in gulpfile.js.  Now we have a common htmlReplace function, called from several places in the build, but it will look for assets with non-fingerprinted names for the various dev builds, and fingerprinted names for a deployment build.